Refs #26029. Create Indirect Plotter
