+++ qt/scientific_interfaces/Indirect/IndirectPlotOptionsPresenter.h
@@ -33,9 +33,6 @@ public:
   void setWorkspaces(std::vector<std::string> const &workspaces);
   void clearWorkspaces();
 
-signals:
-  void runAsPythonScript(QString const &code, bool noOutput = false);
-
 private slots:
   void workspaceChanged(std::string const &workspaceName);
   void indicesChanged(std::string const &indices);
@@ -59,10 +56,6 @@ private:
   void setWorkspace(std::string const &plotWorkspace);
   void setIndices();
 
-#if QT_VERSION < QT_VERSION_CHECK(5, 0, 0)
-  void runPythonCode(boost::optional<std::string> const &plotString);
-#endif
-
   // Observers for ADS Notifications
   Poco::NObserver<IndirectPlotOptionsPresenter,
                   Mantid::API::WorkspacePreDeleteNotification>
@@ -71,10 +64,8 @@ private:
                   Mantid::API::WorkspaceBeforeReplaceNotification>
       m_wsReplacedObserver;
 
-  API::PythonRunner m_pythonRunner;
   std::unique_ptr<IndirectPlotOptionsView> m_view;
   std::unique_ptr<IndirectPlotOptionsModel> m_model;
-  IndirectTab *m_parentTab;
 };
 
 } // namespace CustomInterfaces
